cleanup

+++ ptep_mbs/cmb.py
@@ -82,8 +82,8 @@ class CMBModel:
             freq = self.instr[chnl]['freq']
             if self.params.band_int:
                 band = self.instr[chnl]['freq_band']
-                fmin = int(freq-band/2.)
-                fmax = int(freq+band/2.)
+                fmin = freq-band/2.
+                fmax = freq+band/2.
                 fsteps = int(fmax-fmin+1)
                 bandpass_frequencies = np.linspace(fmin, fmax, fsteps) * u.GHz
                 weights = np.ones(len(bandpass_frequencies))
